TPW_project_2_backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.2.7
django-bootstrap-icons==0.8.6
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
idna==3.4
Markdown==3.5.1
Pillow==10.1.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
